No longer a subclass of Date::Manip::Date
Paradoxically, this turned out to be the best way to make the Date-Manip
test suite pass this distro. The problem was that many pieces of the
Date-Manip interface require an honest-to-God Date::Manip::Date, not a
subclass thereof. Changing the relationship from "isa" to "has-a" was
the easiest way to have such an object to hand to them.

Obviously the Date-Manip test suite had to be munged, but other than a
mass change of class names there were only two:

* The common setup code, t/tests.pl, was changed to hot patch the
  loading of Date::ManipX::Almanac languages so that a failure to load
  becomes a skip_all().

* t/obj.t got patched to instantiate a TZ using $obj->dmd() when it was
  being created from a Date::ManipX::Almanac::Date object.