Tweak tools/convert-to-almanac.
Rather than convert '->_init()' to '->dmd()->_init()', provide an _init()
method in Date::ManipX::Almanac::Date that does the hand-off. The
difference between this and the original (pre- ->dmd()->_init())
implementation is that it is applied by the patched Date::Manip testing
machinery, rather than in Date::ManipX::Almanac::Date, so it is only
present when actually running the patched Date::Manip test suite.